# server/src/main/resources/static/swagger.json
{
	"swagger": "2.0",
	"tags": [
		{
			"name": "device"
		},
		{
			"name": "sysinfo"
		}
	],
	"paths": {
		"/bacnetz/api/device/all": {
			"get": {
				"tags": [
					"device"
				],
				"summary": "Retrieve all devices",
				"description": "",
				"operationId": "all",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bacnetz/api/device/details/{uid}": {
			"get": {
				"tags": [
					"device"
				],
				"summary": "Retrieve detailed device information",
				"description": "",
				"operationId": "details",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "path",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/DeviceDto"
						}
					}
				}
			}
		},
		"/bacnetz/api/device/toggle": {
			"post": {
				"tags": [
					"device"
				],
				"summary": "Toggle all door open states",
				"description": "",
				"operationId": "toggleAll",
				"consumes": [
					"application/json"
				],
				"parameters": [
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/bacnetz/api/device/toggle/{uid}": {
			"post": {
				"tags": [
					"device"
				],
				"summary": "Toggle specific door open state",
				"description": "",
				"operationId": "toggle",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "path",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/bacnetz/api/device/writeProperty": {
			"post": {
				"tags": [
					"device"
				],
				"summary": "Changes a property on a device",
				"description": "",
				"operationId": "toggle_1",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": false,
						"schema": {
							"$ref": "#/definitions/WritePropertyDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/bacnetz/api/sysinfo/version": {
			"get": {
				"tags": [
					"sysinfo"
				],
				"summary": "Get app version.",
				"description": "",
				"operationId": "getImageVersion",
				"parameters": [
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		}
	},
	"definitions": {
		"DeviceDto": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int32"
				},
				"objectType": {
					"type": "string",
					"enum": [
						"ANALOG_INPUT",
						"ANALOG_OUTPUT",
						"ANALOG_VALUE",
						"BINARY_INPUT",
						"DEVICE",
						"FILE",
						"LOOP",
						"MULTI_STATE_INPUT",
						"MULTI_STATE_OUTPUT",
						"NOTIFICATION_CLASS",
						"SCHEDULE",
						"MULTI_STATE_VALUE"
					]
				},
				"name": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"presentValue": {
					"type": "object"
				},
				"states": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"deviceProperties": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/DevicePropertyDto"
					}
				},
				"childDevices": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/DeviceDto"
					}
				}
			}
		},
		"DevicePropertyDto": {
			"type": "object",
			"properties": {
				"key": {
					"type": "integer",
					"format": "int32"
				},
				"name": {
					"type": "string"
				},
				"value": {
					"type": "object"
				}
			}
		},
		"WritePropertyDto": {
			"type": "object",
			"properties": {
				"parentDeviceId": {
					"type": "string"
				},
				"childDeviceId": {
					"type": "string"
				},
				"childObjectType": {
					"type": "string",
					"enum": [
						"ANALOG_INPUT",
						"ANALOG_OUTPUT",
						"ANALOG_VALUE",
						"BINARY_INPUT",
						"DEVICE",
						"FILE",
						"LOOP",
						"MULTI_STATE_INPUT",
						"MULTI_STATE_OUTPUT",
						"NOTIFICATION_CLASS",
						"SCHEDULE",
						"MULTI_STATE_VALUE"
					]
				},
				"propertyKey": {
					"type": "integer",
					"format": "int32"
				},
				"propertyName": {
					"type": "string"
				},
				"value": {
					"type": "object"
				}
			}
		}
	}
}